requirements.txt
torch~=2.2.2
torchvision~=0.17.2
argparse~=1.4.0
Pillow~=9.3.0
setuptools~=49.2.1